requirements.txt
apache-airflow
faker-clickstream